ARM64 HI32: Apply 5-part fix for pointer truncation (partial - still debugging)
Applied comprehensive fixes based on deep analysis documented in ARM64_DEEP_ANALYSIS_HI32_TRUNCATION.md:

**Fix 1: Makefile - Enable HL_JIT_ARM64 flag**
- Added -DHL_JIT_ARM64 to CFLAGS for ARM64/aarch64 architectures
- Ensures T_SIZES and other conditional ARM64 code compiles correctly

**Fix 2: T_SIZES already applied** (src/std/types.c:48-52)
- HI32 allocates 8 bytes on ARM64 (was 4)
- Verified in libhl.so: T_SIZES[3] = 8

**Fix 3: module.c - Zero-extend HI32 constants** (line 601-609)
- Cast m->code->ints[idx] to (unsigned int) before (int_val)
- Prevents sign-extension of 32-bit values to 64-bit
- Critical for pointer values stored in bytecode ints[] array

**Fix 4: jit.c OInt - Zero-extend integer constants** (line 4510-4517)
- Cast m->code->ints[o->p2] to (unsigned int)
- Prevents sign-extension when loading immediate values
- Ensures HI32-typed pointers load correctly as 64-bit

**Fix 5: LOAD_VREG/STORE_VREG already applied** (jit.c:4390, 4413)
- Uses _size=3 (64-bit) for HI32 type
- Matches vreg->size=8 fix at jit.c:3926-3931

**Current Status:**
Program compiles and runs without immediate crash, but still encountering pointer truncation in some code paths (x0=0xf5c433ed when calling hl_hbset_impl). Further investigation needed to identify remaining truncation sources.

**Testing:**
- ./hl --version: ✓ works
- ./hl /tmp/empty.hl: Runs but still has pointer truncation issues
- T_SIZES verified correct in libhl.so

Related to #ARM64 #JIT #Pointer-Truncation

+++ src/jit.c
@@ -4510,7 +4510,8 @@ int_val hl_jit_function( jit_ctx *ctx, hl_module *m, hl_function *f ) {
 	case OInt:
 		// dst = integer constant
 		if (dst) {
-			arm_load_imm64(ctx, X10, m->code->ints[o->p2]);
+			// CRITICAL: Cast to unsigned to zero-extend for HI32 pointers
+			arm_load_imm64(ctx, X10, (unsigned int)m->code->ints[o->p2]);
 			STORE_VREG(X10, dst);
 		}
 		break;
